Move ui component to mobileUi package
- fix fab showing issue
