+++ app/src/main/java/com/kaloglu/sample/mobileui/demo/DemoFragment.kt
@@ -1,4 +1,4 @@
-package com.kaloglu.sample.adapter.demo
+package com.kaloglu.sample.mobileui.demo
 
 import android.os.Bundle
 import android.view.View
