+++ app/src/main/res/layout/fragment_demo.xml
@@ -4,6 +4,11 @@
              android:layout_width="match_parent"
              android:layout_height="match_parent">
 
+    <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text=":))))"/>
+
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recycler_view"
             android:layout_width="match_parent"
